Implement TrinketModifiers.java and re-wire methods, removing previous versions.
