+++ src/main/java/dev/emi/trinkets/api/event/TrinketUnequipCallback.java
@@ -9,11 +9,11 @@ import net.minecraft.item.ItemStack;
 
 public interface TrinketUnequipCallback {
     Event<TrinketUnequipCallback> EVENT = EventFactory.createArrayBacked(TrinketUnequipCallback.class,
-        listeners -> (stack, slot, entity) -> {
-            for (TrinketUnequipCallback listener: listeners){
-                listener.onUnequip(stack, slot, entity);
-            }
-        });
+    listeners -> (stack, slot, entity) -> {
+        for (TrinketUnequipCallback listener: listeners){
+            listener.onUnequip(stack, slot, entity);
+        }
+    });
 
     /**
      * Called when an entity un-equips a trinket, after the {@link Trinket#onUnequip} method of the Trinket
